P2PoW Web Demo
This application is completely client-side and runs directly in the browser. It offers a friendly interface allowing any user to test the P2PoW protocol, using online workers to validate their transactions.

Due to the fact that P2PoW does not support ssl yet, the use of https is not recommended.